Windows: Commonise TSOHandlerConfig
